# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GTCP — Gas Trading & Commercial Platform API
  description: |
    **Gastrans Serbia · Backend REST API**

    Система управления газовыми транзитными потоками, договорами мощности, биллингом и аукционами.

    ### Нормативная база
    - **CAM NC EU 2017/459** — механизмы распределения мощности
    - **АЕРС 05-145** — официальные тарифы Gastrans
    - **Gastrans Network Code** — NC Art. 5 (Credit Support), Art. 6.3.1 (EXIT-SERBIA),
      Art. 8.3 (Surrender), Art. 18 (Fuel Gas), Art. 20 (Billing & Late Payment)
    - **MAR0277-24** — расписание аукционов CAM NC 2025-2026 (ENTSOG)

    ### Аутентификация
    Все endpoints (кроме `/auth/login`) требуют JWT Bearer token.

    ### Lifecycle аукциона
    ```
    Free Capacity → POST /auctions/bids (DRAFT)
                  → POST /auctions/bids/{id}/submit (SUBMITTED)
                  → POST /auctions/bids/{id}/result (WON/LOST)
                  → POST /auctions/bids/{id}/create-contract (CONTRACT_CREATED)
                  → POST /billing (Invoice)
    ```
  version: 5.0.0
  contact:
    name: GTCP Backend
    email: krav4ukleo@gmail.com
  license:
    name: Proprietary — Gastrans Serbia
    url: https://gastrans.rs
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development
  - url: https://gtcp.gastrans.rs/api/v1
    description: Production (Sprint 6)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token, срок действия 24 часа.
        Получить: `POST /auth/login` → `{ token }`.
  schemas:
...
    Contract:
      type: object
      properties:
        id:                     { type: integer }
        contract_number:        { type: string, example: GTA-2026-001 }
        shipper_id:             { type: integer }
        product_type:           { type: string, enum: [ANNUAL, QUARTERLY, MONTHLY, DAILY, WITHIN_DAY] }
        capacity_type:          { type: string, enum: [FIRM, INTERRUPTIBLE] }
        flow_direction:
          type: string
          enum: [GOSPODJINCI_HORGOS, HORGOS_GOSPODJINCI, KIREVO_EXIT_SERBIA]
          description: |
            GOSPODJINCI_HORGOS  = Firm Transit: Kirevo ENTRY → Horgoš EXIT (tariff 4.19+6.85)
            HORGOS_GOSPODJINCI  = Commercial Reverse: Horgoš ENTRY → Kirevo EXIT (3.25)
            KIREVO_EXIT_SERBIA  = Domestic Delivery: Kirevo ENTRY → EXIT-SERBIA (6.00+4.19)
        cap_entry_kwh_h:        { type: number, example: 13752230 }
        cap_exit_kwh_h:         { type: number, example: 9216209 }
        tariff_entry_eur_kwh_h_yr: { type: number, example: 4.19 }
        tariff_exit_eur_kwh_h_yr:  { type: number, example: 6.85 }
        start_date:             { type: string, format: date }
        end_date:               { type: string, format: date }
        status:                 { type: string, enum: [ACTIVE, EXPIRED, PENDING, CANCELLED] }
    ContractCreate:
      type: object
      required: [shipper_id, product_type, flow_direction, cap_entry_kwh_h, cap_exit_kwh_h, start_date, end_date]
      properties:
        shipper_id:         { type: integer }
        product_type:       { type: string, enum: [ANNUAL, QUARTERLY, MONTHLY, DAILY, WITHIN_DAY] }
        capacity_type:      { type: string, enum: [FIRM, INTERRUPTIBLE], default: FIRM }
        flow_direction:
          type: string
          enum: [GOSPODJINCI_HORGOS, HORGOS_GOSPODJINCI, KIREVO_EXIT_SERBIA]
        cap_entry_kwh_h:    { type: number, example: 13752230 }
        cap_exit_kwh_h:     { type: number, example: 9216209 }
        start_date:         { type: string, format: date, example: "2025-10-01" }
        end_date:           { type: string, format: date, example: "2026-10-01" }

    # ─── Nomination ───────────────────────────────────────────
...
security:
  - BearerAuth: []
paths:
...
  /contracts:
    get:
      tags: [contracts]
      summary: Список договоров
      description: Договоры CAM NC — 7 типов (Annual/Quarterly/Monthly/Daily/WithinDay × Firm/Interruptible), 3 направления потока.
      parameters:
        - in: query
          name: shipper_id
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string, enum: [ACTIVE, EXPIRED, PENDING, CANCELLED] }
        - in: query
          name: flow_direction
          schema: { type: string, enum: [GOSPODJINCI_HORGOS, HORGOS_GOSPODJINCI, KIREVO_EXIT_SERBIA] }
        - in: query
          name: product_type
          schema: { type: string, enum: [ANNUAL, QUARTERLY, MONTHLY, DAILY, WITHIN_DAY] }
      responses:
        '200':
          description: Список договоров
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Contract' }

    post:
      tags: [contracts]
      summary: Создать договор
      description: |
        Нумерация GTA-YYYY-NNN. Тарифы АЕРС 05-145 подставляются автоматически по flow_direction.
        Cap entry/exit billing: `fee = cap_entry × t_entry/365 × days + cap_exit × t_exit/365 × days`
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContractCreate' }
      responses:
        '201':
          description: Договор создан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contract' }
  /contracts/{id}:
    get:
      tags: [contracts]
      summary: Детальная карточка договора
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contract' }
    patch:
      tags: [contracts]
      summary: Обновить договор
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [ACTIVE, EXPIRED, PENDING, CANCELLED] }
                cap_entry_kwh_h: { type: number }
                cap_exit_kwh_h:  { type: number }
                notes:           { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contract' }

  # ════════════════════════════════════════════════════════════
  # CAPACITY
  # ════════════════════════════════════════════════════════════
...